--- test/node-api/node-api.status
@@ -1,9 +1,0 @@
-prefix node-api
-
-# To mark a test as flaky, list the test name in the appropriate section
-# below, without ".js", followed by ": PASS,FLAKY". Example:
-# sample-test                        : PASS,FLAKY
-
-[true] # This section applies to all platforms
-
-[$system==win32]
