--- test/node-api/test_instance_data/binding.gyp
@@ -1,10 +1,0 @@
-{
-  "targets": [
-    {
-      "target_name": "test_instance_data",
-      "sources": [
-        "test_instance_data.c"
-      ]
-    }
-  ]
-}
